added new file metadata structure and reorganize tokens and types
