+++ src/@types/index.ts
@@ -3,3 +3,4 @@ export * from './Asset'
 export * from './Service'
 export * from './Credentials'
 export * from './Metadata'
+export * from './FileMetadata'
