--- src/datatokens/index.ts
@@ -1,2 +1,0 @@
-export * from './Datatoken'
-export * from './NFTDatatoken'
